% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tagcloud.R
\name{calc.cex}
\alias{calc.cex}
\title{Calculate cex for tags}
\usage{
calc.cex(x, floor = 1, ceiling = 3, wmin = NULL, wmax = NULL)
}
\arguments{
\item{x}{weights}

\item{floor, ceiling}{minimum and maximum cex calculated (default: 1, 3)}

\item{wmin, wmax}{minimum and maximum weight (default: NULL)}
}
\description{
Calculate cex for tags
}
\details{
Scale the weights such that minimum weight corresponds to minimum font
magnification parameter (cex) and maximum weight corresponds to maximum
cex.
}
