% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/strmultline.R
\name{strmultline}
\alias{strmultline}
\title{Replace some spaces in multi-word sentences by newlines}
\usage{
strmultline(strings, ratio = 0.2)
}
\arguments{
\item{strings}{a character vector containing the multi-word sentences to be
split}

\item{ratio}{the desired ratio height / width}
}
\value{
A character vector containing the modified sentences.
}
\description{
Replace a space character by a newline in a multi-word sentence to get a
better height / width ratio
}
\details{
Very long tags, for example GO Term descriptions, make a bad tag cloud.
\code{strmultline} tries to chop up such a long sentence into multiple
(currently two) lines, to get a better height / width ratio.
}
\author{
January Weiner <january.weiner@gmail.com>
}
\seealso{
\code{\link{tagcloud}}
}
\keyword{splitting}
\keyword{strings}

